requirements.txt
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-mock==3.12.0
python-dateutil==2.8.2